chat_main/requirements.txt
fastapi==0.120.0
httptools==0.7.1
httpx==0.28.1
Jinja2==3.1.6
pyahocorasick==2.2.0
python-dotenv==1.1.1
PyYAML==6.0.3
requests==2.32.5
sqlmodel==0.0.27
uvicorn==0.38.0
watchfiles==1.1.1
websockets==15.0.1